Fix linting for Header
