+++ client/.eslintrc.js
@@ -10,6 +10,7 @@ module.exports = {
       },
     ],
     'react/jsx-filename-extension': [1, { 'extensions': ['.js', '.jsx'] }],
+    'react/prefer-stateless-function': 'off',
   },
   env: {
     browser: true,
